test(workflow): focus activity interruption coverage
